chore: bump version to 1.5.8-beta.3 and establish frontend TDD suite
