+++ custom_components/house_battery_control/manifest.json
@@ -15,5 +15,5 @@
     "scipy>=1.14.0",
     "numpy>=2.1.0"
   ],
-  "version": "1.5.7"
+  "version": "1.5.8-beta.3"
 }
