+++ hacs.json
@@ -6,5 +6,5 @@
   ],
   "iot_class": "local_polling",
   "homeassistant": "2024.1.0",
-  "version": "1.5.7"
+  "version": "1.5.8-beta.3"
 }
