+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "house-battery-control",
+  "version": "1.0.0",
+  "description": "[![HACS](https://img.shields.io/badge/HACS-Custom-41BDF5.svg)](https://hacs.xyz)\r [![HA](https://img.shields.io/badge/Home%20Assistant-2024.1%2B-blue.svg)](https://www.home-assistant.io/)\r [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](LICENSE)",
+  "main": "index.js",
+  "directories": {
+    "doc": "docs",
+    "test": "tests"
+  },
+  "scripts": {
+    "test": "web-test-runner"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/RangeyRover/home-battery-control.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "type": "module",
+  "bugs": {
+    "url": "https://github.com/RangeyRover/home-battery-control/issues"
+  },
+  "homepage": "https://github.com/RangeyRover/home-battery-control#readme",
+  "devDependencies": {
+    "@open-wc/testing": "^4.0.0",
+    "@web/test-runner": "^0.20.2",
+    "sinon": "^21.1.2"
+  }
+}
