+++ specs/038-frontend-refactor/tasks.md
@@ -6,27 +6,27 @@ This document contains the actionable, dependency-ordered tasks to complete the 
 
 The setup phase initializes the file structure required for the new sub-components.
 
-- [ ] T001 Initialize empty LitElement files in `custom_components/house_battery_control/frontend/hbc-dashboard.js`, `frontend/hbc-sensors.js`, and `frontend/hbc-plan-table.js`.
+- [x] T001 Initialize empty LitElement files in `custom_components/house_battery_control/frontend/hbc-dashboard.js`, `frontend/hbc-sensors.js`, and `frontend/hbc-plan-table.js`.
 
 ## Phase 2: Foundational Component Extraction
 
 > **User Story 1**: Decoupling visual components from the orchestrator logic into atomic Web Components strictly constrained to under 300 LOC.
 
-- [ ] T002 [US1] Migrate `_renderDashboard` SVG generation, flow-graph CSS, and HTML from `frontend/hbc-panel.js` into the `render()` method of `frontend/hbc-dashboard.js`, ensuring it registers as `<hbc-dashboard>`.
-- [ ] T003 [P] [US1] Migrate `_renderSensors` layout, error badge logic, and sensor CSS from `frontend/hbc-panel.js` into `frontend/hbc-sensors.js`, ensuring it registers as `<hbc-sensors>`.
-- [ ] T004 [US1] Migrate the comprehensive `_renderPlan` algorithm, mathematically chunking blocks, `_calculateSummaryStats`, and table CSS from `frontend/hbc-panel.js` into `frontend/hbc-plan-table.js`, ensuring it registers as `<hbc-plan-table>`.
+- [x] T002 [US1] Migrate `_renderDashboard` SVG generation, flow-graph CSS, and HTML from `frontend/hbc-panel.js` into the `render()` method of `frontend/hbc-dashboard.js`, ensuring it registers as `<hbc-dashboard>`.
+- [x] T003 [P] [US1] Migrate `_renderSensors` layout, error badge logic, and sensor CSS from `frontend/hbc-panel.js` into `frontend/hbc-sensors.js`, ensuring it registers as `<hbc-sensors>`.
+- [x] T004 [US1] Migrate the comprehensive `_renderPlan` algorithm, mathematically chunking blocks, `_calculateSummaryStats`, and table CSS from `frontend/hbc-panel.js` into `frontend/hbc-plan-table.js`, ensuring it registers as `<hbc-plan-table>`.
 
 ## Phase 3: The Orchestrator Engine
 
 > **User Story 2**: Pushing the data payload to sub-components cleanly without prop-drilling or monolithic render cycles.
 
-- [ ] T005 [US2] Update `frontend/hbc-panel.js` by unconditionally deleting all previously extracted internal rendering/styling methods (`_renderDashboard`, `_renderSensors`, `_renderPlan`, `_calculateSummaryStats`).
-- [ ] T006 [US2] Update `frontend/hbc-panel.js` to utilize ES6 module imports for the new components (`import './hbc-dashboard.js';` etc) and refactor the primary `render()` block to dynamically return the custom tags (`<hbc-dashboard .data=${this._data}></hbc-dashboard>`, etc) strictly driven by `this._activeTab`.
+- [x] T005 [US2] Update `frontend/hbc-panel.js` by unconditionally deleting all previously extracted internal rendering/styling methods (`_renderDashboard`, `_renderSensors`, `_renderPlan`, `_calculateSummaryStats`).
+- [x] T006 [US2] Update `frontend/hbc-panel.js` to utilize ES6 module imports for the new components (`import './hbc-dashboard.js';` etc) and refactor the primary `render()` block to dynamically return the custom tags (`<hbc-dashboard .data=${this._data}></hbc-dashboard>`, etc) strictly driven by `this._activeTab`.
 
 ## Phase 4: Polish & Cross-Cutting Concerns
 
-- [ ] T007 Verify the dynamic tab-reload logic safely initializes all sub-components without browser layout thrashing on cold cache.
-- [ ] T008 Prove `localStorage` keys for the Plan Table hidden columns (`hbc_hidden_cols`) and Sensor states (`hbc_sensors_hidden`) successfully isolate to their respective child components.
+- [x] T007 Verify the dynamic tab-reload logic safely initializes all sub-components without browser layout thrashing on cold cache.
+- [x] T008 Prove `localStorage` keys for the Plan Table hidden columns (`hbc_hidden_cols`) and Sensor states (`hbc_sensors_hidden`) successfully isolate to their respective child components.
 
 ---
