+++ web-test-runner.config.js
@@ -1,0 +1,9 @@
+export default {
+  nodeResolve: true,
+  files: ['tests/js/**/*.test.js'],
+  testFramework: {
+    config: {
+      timeout: 10000,
+    },
+  },
+};
